+++ sparta_constants.go
@@ -18,7 +18,7 @@ const (
 )
 const (
 	// SpartaVersion defines the current Sparta release
-	SpartaVersion = "1.9.2"
+	SpartaVersion = "1.9.3"
 	// GoLambdaVersion is the Go version runtime used for the lambda function
 	GoLambdaVersion = "go1.x"
 	// LambdaBinaryTag is the build tag name used when building the binary
@@ -93,4 +93,7 @@ const (
 	// ContextKeyLambdaResponse is the possible response that
 	// was returned from the lambda function
 	ContextKeyLambdaResponse
+	// ContextKeyAWSSession is the aws Session instance for this
+	// request
+	ContextKeyAWSSession
 )